Deleted two N3DADE example ouput files
The N3DADE legacy source is dated 3/98. The example output files were
apparently generated with the 4/97 version of the code that is listed in
the pdf documentation (no digital version exists as far as I know). In
addition to some changes in output format statements, the 3/98 version
apparently differs in the specification of Dirac BCs.

Line 84 in the 3/98 code is:

DMB=DMB/CO/UL

whereas the 3/97 code is:

DMB=DMB/CO

The example output files ONEA.OUT and ONEB.OUT were not consistent with
the provided 3/98 source, so they were deleted.